refactor(STK-90): rename stacks folder into stack-tabs for meaningfulness
