+++ packages/app/components/index.ts
@@ -1,7 +1,7 @@
 import { ConnectButton } from "./ConnectButton";
 import { Navbar } from "./navbar/Navbar";
 import { SelectNetwork } from "./SelectNetwork";
-import { StackTabs } from "./stacks/StackTabs";
+import { StackTabs } from "./stack-tabs/StackTabs";
 import TokenPicker from "./token-picker/TokenPicker";
 
 export * from "./stackbox";
