+++ packages/app/components/stack-tabs/StackProgress.tsx
